+++ javascript/libjoynr-js/src/main/js/joynr/proxy/ProxyAttributeNotifyRead.js
@@ -86,7 +86,10 @@ define("joynr/proxy/ProxyAttributeNotifyRead", [ "joynr/proxy/ProxyAttribute"
          *            successfully, method signature: "void onReceive({?}value)"
          * @param {Function}
          *            settings.onError this function is called if a publication of the attribute
-         *            value was missed, method signature: "void onError()"
+         *            value was missed, method signature: "void onError({Error} error)"
+         * @param {Function}
+         *            settings.onSubscribed the callback to inform once the subscription request has
+         *            been delivered successfully
          * @returns {Object} returns an A+ promise object. This object additionally holds the unique
          *            subscriptionId, used for unsubscribing from the attribute in the key
          *            subscriptionId